% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{create_data_project}
\alias{create_data_project}
\title{Create a data project stump}
\usage{
create_data_project(name)
}
\arguments{
\item{name}{the name of the project}
}
\description{
Creates all directories and first files for a data project
}
