% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMelanoma-package.R
\docType{package}
\name{dataMelanoma-package}
\alias{dataMelanoma}
\alias{dataMelanoma-package}
\title{dataMelanoma: A data source for melanoma data}
\description{
This package provides various melanoma datasets. It gives access to the code used to harvest/preprocess the data, as well as the data itself. The processed data is stored with git-lfs as long as possible. If github's quota is reached, I might switch to zenodo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrlab/dataMelanoma}
  \item \url{https://ggrlab.github.io/dataMelanoma/}
  \item Report bugs at \url{https://github.com/ggrlab/dataMelanoma/issues}
}

}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
