% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harel2019.R
\docType{data}
\name{harel2019_pheno}
\alias{harel2019_pheno}
\title{Phenodata for Harel2019}
\format{
A data frame with 109 rows and 14 variables. Highlighted columns:
\describe{
\item{\code{Response}}{Response to treatment: PD, PR, CR or SD. Samples with SD were excluded.}
\item{\code{Sample taken after treatment}}{There are 2 samples which were taken AFTER treatment!}
\item{\code{response_paper}}{character COLUMN_DESCRIPTION}
}
}
\source{
Harel et al. "Proteomics of melanoma response to immunotherapy reveals mitochondrial
dependence." Cell 179.1 (2019): 236-250. https://doi.org/10.1016/j.cell.2019.08.012
}
\usage{
harel2019_pheno
}
\description{
Phenodata taken from Supplementary table S1A. Samples have response to
treatment (PD, PR, CR or SD). Stable disease (SD) was excluded from the analysis
following the paper.
}
\details{
In \link{Harel et al, 2019, Cell}, the authors describe melanoma response
to immunotherapy. They use high-resolution mass spectrometry to quantify 4,300
proteins across 116 patients. Formalin-fixed paraffin-embedded (FFPE) tissues from
42 metastatic melanoma patients treated with TIL-based adoptive cell transfer and 74
treated with anti-PD1 immunotherapy were included.

Samples were taken before treatment except for two patients taken 11 or 18 days after
treatment initialization. Patients are split into 21/40 responders,
21/27 non-responders and 0/7 stable disease according to RECIST v1.0
guidelines (TIL/anti-PD1 cohorts). See \verb{Tumor sample collection} in the paper's
method section.
}
\keyword{datasets}
