% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harel2019.R
\docType{data}
\name{harel2019_protmats}
\alias{harel2019_protmats}
\title{Protein matrices for Harel2019}
\format{
A list with two elements:
\describe{
\item{\code{aPD1}}{
Protein matrix with 4620 genes (rows) of 67 samples (columns) from
Supplementary table S1D. Rownames are gene names, columns are sample IDs.
}
\item{\code{til}}{
Protein matrix with 4588 genes (rows) of 42 samples (columns) from
Supplementary table S1C. Rownames are gene names, columns are sample IDs.
}
}
}
\source{
Harel et al. "Proteomics of melanoma response to immunotherapy reveals mitochondrial
dependence." Cell 179.1 (2019): 236-250. https://doi.org/10.1016/j.cell.2019.08.012
}
\usage{
harel2019_protmats
}
\description{
Data taken from Supplementary tables S1C and S1D.
See also \code{\link{harel2019_pheno}}.
}
\details{
We used the data available under \href{https://ars.els-cdn.com/content/image/1-s2.0-S0092867419309006-mmc1.xlsx}{Table S1}
of the publication in \code{1-s2.0-S0092867419309006-mmc1.xlsx},
sheets \code{S1C} and \code{S1D}.

!!Note!! that it seems that the data in Table S2B is inconsistent to S1B, S1D and S4A.

According to the section \verb{Proteomics statistical analysis} in the paper's method
section, the following analysis was performed .

\if{html}{\out{<div class="sourceCode">}}\preformatted{ For the clinical tumor data, normalized ratio tumor/SILAC data were
 log2-transformed and the protein groups were filtered to have at least 70\%
 valid values, reaching a list of 4588,4620 or 4416 protein groups for the
 TIL, the anti-PD1 or the combined datasets respectively, which were further used
 for all downstream analyses. Dataset integration was based on gene name;
 multiple entries for the same gene name were integrated to a single entry by
 calculating the median expression value. Data were normalized by subtracting
 most frequent value in each sample. To extract DEPs between responders and
 non-responders multiple comparisons were performed; a two-sample Student's ttest
 was performed with a p value threshold of 0.05 for each of the two datasets,
 following by 2D annotation enrichment test (Cox and Mann, 2012) with FDR q-value
 0.02 to derive the differential functional groups in each of the two datasets.
 A two-sample Student's t test with a permutation-based FDR q-value of 0.1 and
 S0 (Tusher et al., 2001) of 0.1 was performed for the anti-PD1 dataset or the
 combined dataset of both treatments, to extract significantly changing proteins.
}\if{html}{\out{</div>}}

From main text; \verb{Functional Analysis of Responders and Non-responders to Immunotherapy}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Initial bioinformatic analysis examined the functional differences between
 responders and non-responders in each therapy. We first applied a low stringency
 Student's t test with a nominal p value cutoff (p value < 0.05) and found 414
 and 636 differentially expressed proteins (DEPs) in the TIL and the anti-PD1
 cohorts [...].
}\if{html}{\out{</div>}}
}
\keyword{datasets}
