% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harel2019.R
\docType{data}
\name{harel2019_responder_noSD}
\alias{harel2019_responder_noSD}
\title{Named non-/responder status for Harel2019}
\format{
A named character vector of length 109. Names are the sample IDs (harel2019_pheno[\link{"Sample ID"}]),
values are the response to treatment ("non-responder" or "responder").
}
\source{
Harel et al. "Proteomics of melanoma response to immunotherapy reveals mitochondrial
dependence." Cell 179.1 (2019): 236-250. https://doi.org/10.1016/j.cell.2019.08.012
}
\usage{
harel2019_responder_noSD
}
\description{
Phenodata taken from Supplementary table S1A.
See also \code{\link{harel2019_pheno}}.
}
\keyword{datasets}
