% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee2022_pheno_curated}
\alias{lee2022_pheno_curated}
\title{Lee2022 (curated) phenodata}
\format{
A tibble with 315 rows (samples) and 14 columns (features):

study_name, subject_id, body_site, age, age_category, gender, PMID, number_reads,
number_bases, minimum_read_length, median_read_length, NCBI_accession,
lee_subcohort, PFS12
}
\source{
Lee, K.A., Thomas, A.M., Bolte, L.A. et al. Cross-cohort gut microbiome
associations with immune checkpoint inhibitor response in advanced melanoma. Nat Med
ü28, 535–544 (2022). https://doi.org/10.1038/s41591-022-01695-5
}
\usage{
lee2022_pheno_curated
}
\description{
Manually selected columns from lee2022_treesummarizedexperiment, see
also \code{\link{lee2022_treesummarizedexperiment}}.
}
\keyword{datasets}
