% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee2022_relative_abundance_preprocessed}
\alias{lee2022_relative_abundance_preprocessed}
\title{Lee2022 preprocessed microbiome data}
\format{
Preprocessed data from 315 samples (columns) and 672 species (rows). The rownames
are coded according to different taxonomies. Use the package "mia" to decode them,
but potentially focus on the \code{\link{lee2022_treesummarizedexperiment}}!

k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales|f__Leuconostocaceae|g__Leuconostoc|s__Leuconostoc_garlicum
}
\usage{
lee2022_relative_abundance_preprocessed
}
\description{
Preprocessed microbiome data from Lee2022, see also
\code{\link{lee2022_treesummarizedexperiment}}. All melanoma data from
curatedMetagenomicData was acquired. Those came from the following other \code{study_name}
refering to other papers.
\itemize{
\item FrankelAE_2017
\item GopalakrishnanV_2018
\item LeeKA_2022
\item MatsonV_2018
\item PetersBA_2019
\item WindTT_2020
}

Data preprocessing was according to "Machine learning analysis." in the paper's methods:
\itemize{
\item Excluding samples with low sequencing depth (number_reads must be > 1e6)
}
\enumerate{
\item Remove species with relative abundances with low overall abundance ("1e-4 maximum abundance cutoff")
\item log10 transformed after pseudocount of 1e-5
\item Standardized as z-scores
}
}
\keyword{datasets}
