% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee2022_treesummarizedexperiment}
\alias{lee2022_treesummarizedexperiment}
\title{Lee2022 treesummarizedexperiment microbiome data}
\format{
Preprocessed data from 315 samples (columns) and 672 species (rows). The rownames
are coded according to different taxonomies. Use the package "mia" to decode them.

k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales|f__Leuconostocaceae|g__Leuconostoc|s__Leuconostoc_garlicum
}
\usage{
lee2022_treesummarizedexperiment
}
\description{
Microbiome data from Lee2022, see also
\code{\link{lee2022_treesummarizedexperiment}}. All melanoma data from
curatedMetagenomicData was acquired. Those came from the following other \code{study_name}
refering to other papers.
\itemize{
\item FrankelAE_2017
\item GopalakrishnanV_2018
\item LeeKA_2022
\item MatsonV_2018
\item PetersBA_2019
\item WindTT_2020
}

No further dataprocessing was done.
}
\keyword{datasets}
