% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_li2019.R
\docType{data}
\name{li2019_pheno}
\alias{li2019_pheno}
\title{Li 2019 phenodata}
\format{
A list of 4 with a single tibble each.
\describe{
\item{s1}{78 samples, 8 columns}
\item{s2}{91 samples, 8 columns}
\item{s3_nivo}{394 samples, 10 columns}
\item{s3_ever}{349 samples, 10 columns}
}

The data.frames \code{s1} and \code{s2} have the following columns:
\describe{
\item{\code{Unique_Subject_Identifier}}{Identifier}
\item{\code{Region}}{Sample region}
\item{\code{Derived_Age_at_Consent}}{Age at consent}
\item{\code{Sex}}{Sex. "M", "F". }
\item{\code{Race}}{"WHITE" or "OTHER"}
\item{\code{BRAF_V600E}}{BRAF_V600E mutation status}
\item{\code{OS}}{Overall survival time}
\item{\code{OS_Censor}}{Overall survival status}
}

The data.frames \code{s3_nivo} and \code{s3_ever} have the following columns:
\describe{
\item{\code{Unique_Subject_Identifier}}{Identifier}
\item{\code{Treatment}}{"NIVOLUMAB" or "EVEROLIMUS"}
\item{\code{Region}}{Sample region}
\item{\code{Sex}}{"M" or "F"}
\item{\code{Derived_Age_at_Consent}}{Age at consent}
\item{\code{Race}}{"WHITE" or "OTHER"}
\item{\code{CRF_MSKCC_Risk_Group}}{"FAVORABLE", "INTERMEDIATE" or "POOR"}
\item{\code{Prior antiangiogenic regimens (≥2)}}{}
\item{\code{OS}}{Overall survival time}
\item{\code{OS_Censor}}{Overall survival status}
}
}
\source{
Liu, D., Schilling, B., Liu, D. et al. Integrative molecular and clinical modeling of
clinical outcomes to PD1 blockade in patients with metastatic melanoma.
Nat Med 25, 1916–1927 (2019). https://doi.org/10.1038/s41591-019-0654-5
}
\usage{
li2019_pheno
}
\description{
List of phenodata per study. There are 3 studies with 4 branches:
\describe{
\item{s1}{melanoma (nivolumab, phase1)}
\item{s2}{RCC (nivolumab, phase1)}
\item{s3_nivo}{RCC (nivolumab, phase3)}
\item{s3_ever}{RCC (everolimus, phase3)}
}
Each of the four list elements contains a data frame with samples as rows and
pheno-information as columns.
}
\keyword{datasets}
