% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_li2019.R
\docType{data}
\name{li2019_studies_clearnames}
\alias{li2019_studies_clearnames}
\title{Li 2019 supplement IDs to clearnames}
\format{
A named character vector with 4 elements. Names are the supplement IDs,
values are the clear names.
}
\usage{
li2019_studies_clearnames
}
\description{
The supplemental data s1, s2, s3_nivo and s3_ever have clear names
in the paper. This named vector maps the supplement IDs to the clear names.
}
\keyword{datasets}
