% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_intraassay_sim00_pure_estimate}
\alias{sim_intraassay_sim00_pure_estimate}
\title{Simulated data from intraassay data}
\format{
A data frame with 100 rows (samples) and 143 columns (features).
Of the 143 columns, 141 are cell population counts as features, and "V1" (column 1) is a sample name and "fake_labels" are randomly sampled group labels.
}
\source{
https://git.uni-regensburg.de/paper/2023-Glehr_rROC_code.git
}
\usage{
sim_intraassay_sim00_pure_estimate
}
\description{
Simulated data from intraassay data, with 100 samples. Created within /home/gugl/clonedgit/paper/2023-Glehr_rROC_code
}
\keyword{datasets}
