% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spencer2021_lkt_ppm}
\alias{spencer2021_lkt_ppm}
\title{Spencer 2021 microbiome data, LKT PPM}
\format{
A tibble with 225 rows (features) and 1 + 167 columns (1 feature, 167 samples).
}
\source{
https://github.com/mda-primetr/Spencer_et_al_2021, Christine N. Spencer et al. ,Dietary fiber and probiotics influence the gut microbiome and melanoma immunotherapy response.Science374,1632-1640(2021).DOI:10.1126/science.aaz7015
}
\usage{
spencer2021_lkt_ppm
}
\description{
Relative abundances of Last Known Taxon features (lowest taxon known for a feature,
up to species), in Parts Per Million (PPM). Data for the samples from Spencer 2021
}
\keyword{datasets}
