% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spencer2021_pheno}
\alias{spencer2021_pheno}
\title{Spencer 2021 microbiome data, phenodata}
\format{
A tibble with 167 rows (samples) and 42 columns (features).
}
\source{
Christine N. Spencer et al. ,Dietary fiber and probiotics influence the gut microbiome and melanoma immunotherapy response.Science374,1632-1640(2021).DOI:10.1126/science.aaz7015
}
\usage{
spencer2021_pheno
}
\description{
Data for the samples from Spencer 2021
}
\keyword{datasets}
