% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang2022_Stem.Sig}
\alias{zhang2022_Stem.Sig}
\title{Zhang 2022 Stem.Sig gene signature}
\format{
A character vector of 454 gene names, identified via a single cell analysis of 34
scRNA-seq datasets.
}
\usage{
zhang2022_Stem.Sig
}
\description{
Stem.Sig gene signature from Zhang2022, see also \code{\link{zhang2022}}.
}
\keyword{datasets}
