% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang2022_test}
\alias{zhang2022_test}
\title{Zhang 2022 test bulk RNAseq data}
\format{
A tibble with 149 rows and 15904 columns. Of those columns, the first 8 are phenodata
per sample (rows): batch, ID, OS, status, PFS, status2, response and BOR. The remaining
columns are gene expression data.
}
\usage{
zhang2022_test
}
\description{
Test data from Zhang2022, post ComBat() preprocessing. See also \code{\link{zhang2022}}.
}
\keyword{datasets}
