% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang2022_training}
\alias{zhang2022_training}
\title{Zhang 2022 training bulk RNAseq data}
\format{
A tibble with 618 rows and 15904 columns. Of those columns, the first 8 are phenodata
per sample (rows): batch, ID, OS, status, PFS, status2, response and BOR. The remaining
columns are gene expression data.
}
\usage{
zhang2022_training
}
\description{
Training data from Zhang2022, post ComBat() preprocessing. See also \code{\link{zhang2022}}.
}
\keyword{datasets}
